refactor: rename ui-heading to ui-heading-legacy
Clear the ui-heading namespace for modern rebuild implementation:
- Rename directory: ui-heading/ → ui-heading-legacy/
- Update component registration: 'ui-heading' → 'ui-heading-legacy'
- Update CSS selectors to ui-heading-legacy
- Update test expectations
- Update main.ts demo to use legacy component

✅ All tests passing (337 tests)
✅ Build successful (28.92 kB)
✅ ui-heading name now available for new implementation
