--- src/components/primitives/ui-heading/ui-heading.css
@@ -1,72 +1,0 @@
-/**
- * UIHeading Component Styles
- * 
- * Uses CSS custom properties for easy theming.
- * Consumers can override any --ui-heading-* variable to customize appearance.
- */
-
-ui-heading {
-  /* Base heading styles */
-  display: block;
-  font-family: var(--ui-heading-font-family, inherit);
-  font-weight: var(--ui-heading-font-weight, 600);
-  line-height: var(--ui-heading-line-height, 1.2);
-  color: var(--ui-heading-color, inherit);
-  margin: var(--ui-heading-margin, 0 0 1rem 0);
-
-  /* No default font-size - level attribute is required */
-  font-size: var(--ui-heading-font-size, 1rem);
-}
-
-/* Level-specific font sizes */
-ui-heading[level='1'] {
-  font-size: var(--ui-heading-h1-size, 2.5rem);
-  font-weight: var(--ui-heading-h1-weight, var(--ui-heading-font-weight, 700));
-}
-
-ui-heading[level='2'] {
-  font-size: var(--ui-heading-h2-size, 2rem);
-  font-weight: var(--ui-heading-h2-weight, var(--ui-heading-font-weight, 600));
-}
-
-ui-heading[level='3'] {
-  font-size: var(--ui-heading-h3-size, 1.5rem);
-  font-weight: var(--ui-heading-h3-weight, var(--ui-heading-font-weight, 600));
-}
-
-ui-heading[level='4'] {
-  font-size: var(--ui-heading-h4-size, 1.25rem);
-  font-weight: var(--ui-heading-h4-weight, var(--ui-heading-font-weight, 600));
-}
-
-ui-heading[level='5'] {
-  font-size: var(--ui-heading-h5-size, 1.125rem);
-  font-weight: var(--ui-heading-h5-weight, var(--ui-heading-font-weight, 600));
-}
-
-ui-heading[level='6'] {
-  font-size: var(--ui-heading-h6-size, 1rem);
-  font-weight: var(--ui-heading-h6-weight, var(--ui-heading-font-weight, 600));
-}
-
-/* 
- * Theme integration
- * These can be overridden in the design system's main CSS file
- */
-
-/* Example: Brand theming
-:root {
-  --ui-heading-font-family: 'Inter', system-ui, sans-serif;
-  --ui-heading-color: var(--color-neutral-900);
-  --ui-heading-h1-size: 3rem;
-  --ui-heading-h2-size: 2.25rem;
-}
-*/
-
-/* Example: Dark mode support
-@media (prefers-color-scheme: dark) {
-  ui-heading {
-    --ui-heading-color: var(--color-neutral-100);
-  }
-}
-*/
