+++ src/main.ts
@@ -5,7 +5,7 @@
 import './style.css';
 
 // Import our components
-import './components/primitives/ui-heading/ui-heading.js';
+import './components/primitives/ui-heading-legacy/ui-heading.js';
 import './components/primitives/ui-button/ui-button.js';
 import type { UIButtonClickEventDetail } from './components/primitives/ui-button/ui-button.js';
 
@@ -16,17 +16,17 @@ document.querySelector<HTMLDivElement>('#app')!.innerHTML = `
     <p>Development playground for testing components</p>
     
     <section id="component-demo">
-      <h2>Minimal UI-Heading Demo</h2>
+      <h2>Legacy UI-Heading Demo</h2>
       
       <!-- Basic headings with semantic levels -->
-      <ui-heading level="1">Heading Level 1</ui-heading>
-      <ui-heading level="2">Heading Level 2</ui-heading>
-      <ui-heading level="3">Heading Level 3</ui-heading>
+      <ui-heading-legacy level="1">Heading Level 1</ui-heading-legacy>
+      <ui-heading-legacy level="2">Heading Level 2</ui-heading-legacy>
+      <ui-heading-legacy level="3">Heading Level 3</ui-heading-legacy>
       
       <!-- Default headings (no variants - keeping it minimal) -->
-      <ui-heading level="4">Heading Level 4</ui-heading>
-      <ui-heading level="5">Heading Level 5</ui-heading>
-      <ui-heading level="6">Heading Level 6</ui-heading>
+      <ui-heading-legacy level="4">Heading Level 4</ui-heading-legacy>
+      <ui-heading-legacy level="5">Heading Level 5</ui-heading-legacy>
+      <ui-heading-legacy level="6">Heading Level 6</ui-heading-legacy>
     </section>
 
     <section id="button-demo">
